+++ kclvm/tools/src/LSP/src/quick_fix.rs
@@ -158,21 +158,16 @@ pub(crate) fn convert_code_to_kcl_diag_id(code: &NumberOrString) -> Option<Diagn
 
 #[cfg(test)]
 mod tests {
-
-    use kclvm_driver::toolchain;
     use lsp_types::{
         CodeAction, CodeActionKind, CodeActionOrCommand, Diagnostic, Position, Range, TextEdit,
         Url, WorkspaceEdit,
     };
-    use parking_lot::RwLock;
     use proc_macro_crate::bench_test;
-    use std::{path::PathBuf, sync::Arc};
+    use std::path::PathBuf;
 
     use super::quick_fix;
     use crate::{
-        state::KCLVfs,
-        to_lsp::kcl_diag_to_lsp_diags_by_file,
-        util::{compile_with_params, Params},
+        state::KCLVfs, tests::kcl_diag_to_lsp_diags_by_file, util::{compile_with_params, Params}
     };
 
     #[test]
@@ -188,8 +183,6 @@ mod tests {
             module_cache: None,
             scope_cache: None,
             vfs: Some(KCLVfs::default()),
-            entry_cache: None,
-            tool: Arc::new(RwLock::new(toolchain::default())),
             gs_cache: None,
         })
         .0;
