+++ kclvm/tools/src/LSP/src/tests.rs
@@ -2,6 +2,7 @@ use crossbeam_channel::after;
 use crossbeam_channel::select;
 use indexmap::IndexSet;
 use kclvm_ast::MAIN_PKG;
+use kclvm_driver::lookup_compile_workspace;
 use kclvm_driver::toolchain;
 use kclvm_driver::toolchain::Metadata;
 use kclvm_driver::WorkSpaceKind;
@@ -53,7 +54,6 @@ use std::process::Command;
 use std::sync::Arc;
 use std::thread;
 use std::time::Duration;
-use std::time::Instant;
 
 use kclvm_ast::ast::Program;
 use kclvm_error::Diagnostic as KCLDiagnostic;
@@ -77,11 +77,11 @@ use crate::from_lsp::file_path_from_url;
 use crate::goto_def::goto_def;
 use crate::hover::hover;
 use crate::main_loop::main_loop;
-use crate::state::KCLEntryCache;
 use crate::state::KCLGlobalStateCache;
 use crate::state::KCLVfs;
-use crate::to_lsp::kcl_diag_to_lsp_diags_by_file;
-use crate::util::lookup_compile_unit_with_cache;
+use crate::to_lsp::kcl_diag_id_to_lsp_diag_code;
+use crate::to_lsp::kcl_err_level_to_severity;
+use crate::to_lsp::kcl_msg_to_lsp_diags;
 use crate::util::to_json;
 use crate::util::{apply_document_changes, compile_with_params, Params};
 
@@ -134,12 +134,36 @@ pub(crate) fn compile_test_file(
         module_cache: Some(KCLModuleCache::default()),
         scope_cache: Some(KCLScopeCache::default()),
         vfs: Some(KCLVfs::default()),
-        entry_cache: Some(KCLEntryCache::default()),
-        tool: Arc::new(RwLock::new(toolchain::default())),
         gs_cache: Some(KCLGlobalStateCache::default()),
     });
     let (program, gs) = compile_res.unwrap();
     (file, program, diags, gs)
+}
+
+/// Convert KCL Diagnostic to LSP Diagnostics.
+pub fn kcl_diag_to_lsp_diags_by_file(diag: &KCLDiagnostic, file_name: &str) -> Vec<Diagnostic> {
+    let mut diags = vec![];
+    for (idx, msg) in diag.messages.iter().enumerate() {
+        if msg.range.0.filename == file_name {
+            let mut related_msg = diag.messages.clone();
+            related_msg.remove(idx);
+            let code = if diag.code.is_some() {
+                Some(kcl_diag_id_to_lsp_diag_code(diag.code.clone().unwrap()))
+            } else {
+                None
+            };
+
+            let lsp_diag = kcl_msg_to_lsp_diags(
+                msg,
+                kcl_err_level_to_severity(diag.level),
+                related_msg,
+                code,
+            );
+
+            diags.push(lsp_diag);
+        }
+    }
+    diags
 }
 
 pub(crate) fn compile_test_file_and_metadata(
@@ -157,22 +181,15 @@ pub(crate) fn compile_test_file_and_metadata(
 
     let file = test_file.to_str().unwrap().to_string();
 
-    let entry_cache = KCLEntryCache::default();
+    let metadata = lookup_compile_workspace(&toolchain::default(), &file, true).2;
     let (diags, compile_res) = compile_with_params(Params {
         file: file.clone(),
         module_cache: Some(KCLModuleCache::default()),
         scope_cache: Some(KCLScopeCache::default()),
         vfs: Some(KCLVfs::default()),
-        entry_cache: Some(entry_cache.clone()),
-        tool: Arc::new(RwLock::new(toolchain::default())),
         gs_cache: Some(KCLGlobalStateCache::default()),
     });
     let (program, gs) = compile_res.unwrap();
-
-    let metadata = entry_cache
-        .read()
-        .get(&file)
-        .and_then(|metadata| metadata.0 .2.clone());
 
     (file, program, diags, gs, metadata)
 }
@@ -333,8 +350,6 @@ fn diagnostics_test() {
         module_cache: None,
         scope_cache: None,
         vfs: Some(KCLVfs::default()),
-        entry_cache: Some(KCLEntryCache::default()),
-        tool: Arc::new(RwLock::new(toolchain::default())),
         gs_cache: Some(KCLGlobalStateCache::default()),
     })
     .0;
@@ -524,8 +539,6 @@ fn complete_import_external_file_test() {
         module_cache: None,
         scope_cache: None,
         vfs: Some(KCLVfs::default()),
-        entry_cache: Some(KCLEntryCache::default()),
-        tool: Arc::new(RwLock::new(toolchain::default())),
         gs_cache: Some(KCLGlobalStateCache::default()),
     })
     .1
@@ -586,8 +599,6 @@ fn goto_import_external_file_test() {
         module_cache: None,
         scope_cache: None,
         vfs: Some(KCLVfs::default()),
-        entry_cache: Some(KCLEntryCache::default()),
-        tool: Arc::new(RwLock::new(toolchain::default())),
         gs_cache: Some(KCLGlobalStateCache::default()),
     });
     let gs = compile_res.unwrap().1;
@@ -939,29 +950,6 @@ fn cancel_test() {
     });
 
     assert!(server.receive_response(id.into()).is_none());
-}
-
-#[test]
-fn entry_test() {
-    let root = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-    let mut path = root.clone();
-    path.push("src/test_data/compile_unit/b.k");
-
-    let path = path.to_str().unwrap();
-
-    let tool = toolchain::default();
-    let entry = KCLEntryCache::default();
-    let start = Instant::now();
-    let _ = lookup_compile_unit_with_cache(&tool, &Some(Arc::clone(&entry)), path);
-
-    assert!(entry.read().get(&path.to_string()).is_some());
-    let first_compile_time = start.elapsed();
-
-    let start = Instant::now();
-    let _ = lookup_compile_unit_with_cache(&tool, &Some(entry), path);
-    let second_compile_time = start.elapsed();
-
-    assert!(first_compile_time > second_compile_time);
 }
 
 #[test]
@@ -1446,8 +1434,6 @@ fn konfig_goto_def_test_base() {
         module_cache: None,
         scope_cache: None,
         vfs: Some(KCLVfs::default()),
-        entry_cache: Some(KCLEntryCache::default()),
-        tool: Arc::new(RwLock::new(toolchain::default())),
         gs_cache: Some(KCLGlobalStateCache::default()),
     })
     .1
@@ -1541,8 +1527,6 @@ fn konfig_goto_def_test_main() {
         module_cache: None,
         scope_cache: None,
         vfs: Some(KCLVfs::default()),
-        entry_cache: Some(KCLEntryCache::default()),
-        tool: Arc::new(RwLock::new(toolchain::default())),
         gs_cache: Some(KCLGlobalStateCache::default()),
     })
     .1
@@ -1608,8 +1592,6 @@ fn konfig_completion_test_main() {
         module_cache: None,
         scope_cache: None,
         vfs: Some(KCLVfs::default()),
-        entry_cache: Some(KCLEntryCache::default()),
-        tool: Arc::new(RwLock::new(toolchain::default())),
         gs_cache: Some(KCLGlobalStateCache::default()),
     })
     .1
@@ -1724,8 +1706,6 @@ fn konfig_hover_test_main() {
         module_cache: None,
         scope_cache: None,
         vfs: Some(KCLVfs::default()),
-        entry_cache: Some(KCLEntryCache::default()),
-        tool: Arc::new(RwLock::new(toolchain::default())),
         gs_cache: Some(KCLGlobalStateCache::default()),
     })
     .1
@@ -2160,8 +2140,6 @@ fn compile_unit_test() {
         module_cache: None,
         scope_cache: None,
         vfs: Some(KCLVfs::default()),
-        entry_cache: Some(KCLEntryCache::default()),
-        tool: Arc::new(RwLock::new(toolchain::default())),
         gs_cache: Some(KCLGlobalStateCache::default()),
     })
     .1
